+++ contracts/interfaces/ICurveRewardsClaimer.sol
@@ -14,7 +14,9 @@ library DataTypes {
 }
 
 interface ICurveRewardsClaimer {
-    function reward_data(address reward) external view returns(DataTypes.RewardToken memory);
-    function last_update_time() external view returns(uint);
+    function reward_data(address reward) external view returns (DataTypes.RewardToken memory);
+
+    function last_update_time() external view returns (uint256);
+
     function get_reward() external;
 }
