+++ contracts/interfaces/ICurveRewardsGaugeReadOnly.sol
@@ -2,5 +2,5 @@
 pragma solidity ^0.7.0;
 
 interface ICurveRewardsGaugeReadOnly {
-    function claimable_reward_write(address user, address token) external view returns(uint);
+    function claimable_reward_write(address user, address token) external view returns (uint256);
 }
