+++ contracts/interfaces/IElevenQuickStrat.sol
@@ -2,6 +2,7 @@
 pragma solidity ^0.7.0;
 
 interface IElevenQuickStrat {
-    function WITHDRAWAL_FEE() external view returns (uint);
-    function WITHDRAWAL_MAX() external view returns (uint);
+    function WITHDRAWAL_FEE() external view returns (uint256);
+
+    function WITHDRAWAL_MAX() external view returns (uint256);
 }
