+++ contracts/interfaces/ISnowballVoter.sol
@@ -4,9 +4,9 @@ pragma solidity ^0.7.0;
 interface ISnowballVoter {
     function execute(
         address to,
-        uint value,
+        uint256 value,
         bytes calldata data
     ) external returns (bool, bytes memory);
 
-    function increaseAmount(uint) external;
+    function increaseAmount(uint256) external;
 }
